finished: 202503 Rust

+++ AdventOfCode_2025_Rust/src/day03/mod.rs
@@ -5,6 +5,6 @@ pub fn solve_part1() -> i32 {
     day03_part1::solve()
 }
 
-pub fn solve_part2() -> i32 {
+pub fn solve_part2() -> i64 {
     day03_part2::solve()
 }
